+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 3. Fixed merged scs feature
 4. Added animation to theme toggler
 5. Added Tracking Area Code
+6. Added New Feature: Connection Monitoring Alerts. This feature sends an automated email once there was a disconnect happened.